Fix Event Signup block form-content appender inheriting frozen preview
editor.js portals the "Form content" InnerBlocks area into a slot render.php
places inside .fair-events-get-tickets so its on-canvas position matches the
frontend. That wrapper carries a pointer-events: none rule to freeze the live
ticket-purchase preview in the editor, which the portaled appender button and
any already-nested question blocks inherited too, making them unclickable.

Narrow the rule with a higher-specificity override scoped to the questions
slot, instead of touching editor.js/render.php's portal structure.

Move the e2e coverage to the root e2e/ harness, which seeds a real event date
so render.php actually nests the slot inside .fair-events-get-tickets. The
retired fair-events/e2e spec opened a blank post-new.php, where no event date
resolves and the slot renders as a plain sibling instead — so it could never
reproduce this bug.

Closes #1344
