+++ .changeset/fix-signup-form-content-appender.md
@@ -1,0 +1,5 @@
+---
+"fair-events": patch
+---
+
+Fix the Event Signup block's "Form content" area being unclickable in the block editor (the add-block appender, and any already-nested question blocks, inherited the live preview's non-interactive styling).
